Created possibility to run on push
